backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
python-dotenv==1.0.1
yfinance==0.2.40
google-genai>=0.8.0
apscheduler==3.10.4
pandas==2.2.2
numpy==1.26.4
pandas-ta==0.3.14b
requests==2.32.3
beautifulsoup4==4.12.3
